requirements.txt
soundfile==0.10.3.post1
librosa==0.8.1
numpy
ffmpeg
audiosegment
tqdm
pandas
pydub==0.24.1
pyyaml
sox
youtube_dl
typeguard
pathlib
requests
torchaudio
speechbrain
ruamel.yaml
pycld3